+++ src/main/kotlin/geet/commands/plumbing/geetWriteTree.kt
@@ -1,7 +1,7 @@
 package geet.commands.plumbing
 
 import geet.exceptions.BadRequest
-import geet.utils.commandutil.writeTree
+import geet.utils.commandutil.plumbingutil.writeTree
 
 fun geetWriteTree(commandLines: Array<String>): Unit {
     if (commandLines.size > 1) {
