+++ src/main/kotlin/geet/utils/commandutil/plumbingutil/updateIndexUtil.kt
@@ -1,4 +1,4 @@
-package geet.utils.commandutil
+package geet.utils.commandutil.plumbingutil
 
 import geet.commands.plumbing.GeetUpdateIndexOptions
 import geet.exceptions.NotFound
