+++ src/main/kotlin/geet/utils/commandutil/plumbingutil/writeTreeUtil.kt
@@ -1,4 +1,4 @@
-package geet.utils.commandutil
+package geet.utils.commandutil.plumbingutil
 
 import geet.managers.IndexData
 import geet.objects.GeetObject
